Moves relationship access out of the LegacyKernelOperations.
 o Still delegates to old implementation, but this is now done through the correct call paths.
 o Locking has been moved to the Kernel locking layer.
 o Obviates the need for NodeLookup#lookup(), which has thus been removed.
 o Obviates the need for the lock that could be passed to getNodeForProxy, which has been removed.
 o Deleted relationships are now filtered out in StateHandlingStatementOperations.
 o Use primitive iterator on the low levels for rels

+++ community/kernel/src/main/java/org/neo4j/kernel/api/DataWrite.java
@@ -31,7 +31,7 @@ interface DataWrite
 
     void nodeDelete( long nodeId );
 
-    long relationshipCreate( long relationshipTypeId, long startNodeId, long endNodeId )
+    long relationshipCreate( int relationshipTypeId, long startNodeId, long endNodeId )
             throws RelationshipTypeIdNotFoundKernelException, EntityNotFoundException;
 
     void relationshipDelete( long relationshipId );
